Bugado mas agora é do Ícaro

+++ controlador/interpolacao/funcoesInterpolacao.cpp
@@ -1,6 +1,6 @@
 #include "interpolacao.h"
 
-void funcaoInterpNewton (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, int grau, expre::expre *funcao, double inter[2], int filosofia, int particoes) {
+void funcaoInterpNewton (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, int grau, expre::expre *funcao, float inter[2], int filosofia, int particoes) {
     pontosInterp.clear();
     pontosFuncao.clear();
 
@@ -39,7 +39,7 @@ void funcaoInterpNewton (std::vector <interp> &pontosInterp, std::vector <interp
     }
 }
 
-void interpNewtonG1Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG1Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -79,7 +79,7 @@ void interpNewtonG1Fechada (std::vector <interp> &pontosInterp, std::vector <int
     }
 }
 
-void interpNewtonG1Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG1Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -126,7 +126,7 @@ void interpNewtonG1Aberta (std::vector <interp> &pontosInterp, std::vector <inte
     }
 }
 
-void interpNewtonG2Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG2Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -168,7 +168,7 @@ void interpNewtonG2Fechada (std::vector <interp> &pontosInterp, std::vector <int
     }
 }
 
-void interpNewtonG2Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG2Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -218,7 +218,7 @@ void interpNewtonG2Aberta (std::vector <interp> &pontosInterp, std::vector <inte
     }
 }
 
-void interpNewtonG3Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG3Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -262,7 +262,7 @@ void interpNewtonG3Fechada (std::vector <interp> &pontosInterp, std::vector <int
     }
 }
 
-void interpNewtonG3Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG3Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, s3, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -314,7 +314,7 @@ void interpNewtonG3Aberta (std::vector <interp> &pontosInterp, std::vector <inte
     }
 }
 
-void interpNewtonG4Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG4Fechada (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
 
@@ -360,7 +360,7 @@ void interpNewtonG4Fechada (std::vector <interp> &pontosInterp, std::vector <int
     }
 }
 
-void interpNewtonG4Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, double inter[2], int particoes) {
+void interpNewtonG4Aberta (std::vector <interp> &pontosInterp, std::vector <interp> &pontosFuncao, expre::expre *funcao, float inter[2], int particoes) {
     double pos, s, s2, s3, s4, dist, passo = QTD_DIST_UM / particoes, fim = inter[1] + QTD_DIST_UM;
     int qtdPontos, tamParticoes, i, j, k;
